navigation-api/navigation-methods/return-value/back-204-205-download.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script src="/common/utils.js"></script>

<body>
<script>
promise_test(async t => {
  const id = token();

  const i = document.createElement("iframe");
  i.src = `resources/204-205-download-on-second-visit.py?id=${id}`;
  document.body.append(i);
  await new Promise(r => i.onload = r);

  // Configure it to return a 204 on the next visit
  await fetch(i.src + "&action=204", { method: "POST" });

  // Now navigate elsewhere
  i.contentWindow.location.href = "/common/blank.html";
  await new Promise(r => i.onload = r);

  // Now try going back. It should do nothing (and not tell us about the result).

  const indexBefore = i.contentWindow.navigation.currentEntry.index;

  const result = i.contentWindow.navigation.back();

  assertNeverSettles(t, result, i.contentWindow);
  i.contentWindow.navigation.onnavigate = t.unreached_func("onnavigate should not be called");
  i.contentWindow.navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  i.contentWindow.navigation.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  await new Promise(resolve => t.step_timeout(resolve, 30));
  assert_equals(i.contentWindow.navigation.currentEntry.index, indexBefore);
}, "back() promises to 204s never settle");

promise_test(async t => {
  const id = token();

  const i = document.createElement("iframe");
  i.src = `resources/204-205-download-on-second-visit.py?id=${id}`;
  document.body.append(i);
  await new Promise(r => i.onload = r);

  // Configure it to return a 205 on the next visit
  await fetch(i.src + "&action=205", { method: "POST" });

  // Now navigate elsewhere
  i.contentWindow.location.href = "/common/blank.html";
  await new Promise(r => i.onload = r);

  // Now try going back. It should do nothing (and not tell us about the result).

  const indexBefore = i.contentWindow.navigation.currentEntry.index;

  const result = i.contentWindow.navigation.back();

  assertNeverSettles(t, result, i.contentWindow);
  i.contentWindow.navigation.onnavigate = t.unreached_func("onnavigate should not be called");
  i.contentWindow.navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  i.contentWindow.navigation.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  await new Promise(resolve => t.step_timeout(resolve, 30));
  assert_equals(i.contentWindow.navigation.currentEntry.index, indexBefore);
}, "back() promises to 205s never settle");

promise_test(async t => {
  const id = token();

  const i = document.createElement("iframe");
  i.src = `resources/204-205-download-on-second-visit.py?id=${id}`;
  document.body.append(i);
  await new Promise(r => i.onload = r);

  // Configure it to return a download on the next visit
  await fetch(i.src + "&action=download", { method: "POST" });

  // Now navigate elsewhere
  i.contentWindow.location.href = "/common/blank.html";
  await new Promise(r => i.onload = r);

  // Now try going back. It should do nothing (and not tell us about the result).

  const indexBefore = i.contentWindow.navigation.currentEntry.index;

  const result = i.contentWindow.navigation.back();

  assertNeverSettles(t, result, i.contentWindow);
  i.contentWindow.navigation.onnavigate = t.unreached_func("onnavigate should not be called");
  i.contentWindow.navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  i.contentWindow.navigation.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  await new Promise(resolve => t.step_timeout(resolve, 30));
  assert_equals(i.contentWindow.navigation.currentEntry.index, indexBefore);
}, "back() promises to Content-Disposition: attachment responses never settle");
</script>
